Additional fixes for generic dialog test utils move
