+++ gradle.properties
@@ -3,7 +3,6 @@ org.gradle.parallel=true
 org.gradle.configureondemand=false
 org.gradle.caching=true
 android.useAndroidX=true
-android.testConfig.useRelativePath=true
 android.enableJetifier=true
 android.enableR8.fullMode=true
 android.defaults.buildfeatures.buildconfig=false
