+++ ime/addons/src/debug/res/values/test_resources.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<resources>
+<resources xmlns:tools="http://schemas.android.com/tools" xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2" tools:ignore="MissingTranslation">
     <string name="test_default_test_addon_id">8774f99e-fb4a-49fa-b8d0-4083f762250a</string>
     <string name="addon_1_description">add-on 1 description</string>
     <string name="addon_1_name">add-on 1</string>
@@ -10,6 +10,4 @@
     <string name="addon_4_description">add-on 4 description</string>
     <string name="addon_4_name">add-on 4</string>
     <integer name="data_2">22</integer>
-
-    <string name="testing_search_title">Search this</string>
 </resources>
