+++ ime/addons/src/main/res/drawable-hdpi/ic_action_search.png
