+++ ime/addons/src/main/res/drawable-xxhdpi/ic_action_search.png
