+++ ime/remote/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
-<resources>
+<?xml version="1.0" encoding="utf-8"?>
+<resources xmlns:tools="http://schemas.android.com/tools" xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2" tools:ignore="MissingTranslation">
     <string name="media_pick_chooser_title">Pick image</string>
 </resources>
